+++ examples/partitioning.cpp
@@ -2,10 +2,6 @@
 
 #include "set_backend.hpp"
 
-namespace bulk {
-using namespace experimental;
-}
-
 int main() {
     environment env;