+++ include/bulk/partitionings/cyclic.hpp
@@ -1,6 +1,6 @@
 #include "partitioning.hpp"
 
-namespace bulk::experimental {
+namespace bulk {
 
 /**
  * A cyclic partitioning distributes the indices of a D-dimension space over the
@@ -71,4 +71,4 @@ class cyclic_partitioning : public multi_partitioning<D, G> {
     }
 };
 
-} // namespace bulk::experimental
+} // namespace bulk