+++ include/bulk/partitionings/partitioning.hpp
@@ -9,7 +9,7 @@
 #include "../util/binary_tree.hpp"
 #include "../util/indices.hpp"
 
-namespace bulk::experimental {
+namespace bulk {
 
 template <int D>
 using index_type = std::array<int, D>;
@@ -115,4 +115,4 @@ class rectangular_partitioning : public multi_partitioning<D, G> {
     }
 };
 
-}  // namespace bulk::experimental
+} // namespace bulk