+++ include/bulk/partitionings/tree.hpp
@@ -1,6 +1,6 @@
 #include "partitioning.hpp"
 
-namespace bulk::experimental {
+namespace bulk {
 
 /**
  * A binary-space partitioning (aptly abbreviated BSP, of course because of its
@@ -104,4 +104,4 @@ class tree_partitioning : public rectangular_partitioning<D, 1> {
     std::vector<index_type<D>> extents_;
 };
 
-} // namespace bulk::experimental
+} // namespace bulk